Fixed some warning.

+++ src/ds5/ds5-auto-calibration.cpp
@@ -400,7 +400,7 @@ namespace librealsense
                     if (progress_callback)
                     {
                         if (host_assistance)
-                            if (count < 20) progress_callback->on_update_progress(80 + count++);
+                            if (count < 20) progress_callback->on_update_progress(static_cast<float>(80 + count++));
                         else
                             progress_callback->on_update_progress(count++ * (2.f * speed)); //curently this number does not reflect the actual progress
                     }
@@ -516,7 +516,7 @@ namespace librealsense
                     if (progress_callback)
                     {
                         if (host_assistance)
-                            if (count < 20) progress_callback->on_update_progress(80 + count++);
+                            if (count < 20) progress_callback->on_update_progress(static_cast<float>(80 + count++));
                         else
                             progress_callback->on_update_progress(count++* (2.f * 3)); //curently this number does not reflect the actual progress
                     }
@@ -637,7 +637,7 @@ namespace librealsense
                         if (progress_callback)
                         {
                             if (host_assistance)
-                                if (count < 20) progress_callback->on_update_progress(80 + count++);
+                                if (count < 20) progress_callback->on_update_progress(static_cast<float>(80 + count++));
                             else
                                 progress_callback->on_update_progress(count++* (2.f * speed)); //curently this number does not reflect the actual progress
                         }
@@ -785,7 +785,7 @@ namespace librealsense
                     if (progress_callback)
                     {
                         if (depth < 0 && count < 20)
-                            progress_callback->on_update_progress(80 + count++);
+                            progress_callback->on_update_progress(static_cast<float>(80 + count++));
                         else if (depth == 0)
                             progress_callback->on_update_progress(count++ * (2.f * speed)); //curently this number does not reflect the actual progress
                     }
